+++ src/core/shared/enums/vehicleTypeFlags.ts
