+++ src/core/shared/interfaces/vehicleInfo.ts
@@ -1,4 +1,4 @@
-import { FUEL_TYPE, VEHICLE_CLASS, VEHICLE_TYPE } from '../enums/vehicleInformationFlags';
+import { FUEL_TYPE, VEHICLE_CLASS, VEHICLE_TYPE } from '../enums/vehicleTypeFlags';
 
 /**
  * Interface for working with / using vehicles.
